Updated shaders to work on NVidia RTX2060 GPU and fixed several projects for Windows

+++ Projects/CastleCreator/Shaders/fs1.glsl
@@ -1,4 +1,5 @@
-precision mediump float;       	// Set the default precision to medium
+precision highp float;       	// Set the default precision to medium
+//Note high precision is needed to NVidia RTX2060 card
 
 uniform sampler2D u_Texture;   	// texture
 varying vec2 v_UV;				// Texture UV coordinate
