+++ Projects/ShaderToy/Shaders/Atmosphere.fs
@@ -2,7 +2,7 @@
 // Written by GLtracy
 
 #version 300 es
-precision mediump float;       	// Set the default precision to medium
+precision highp float;       	// Set the default precision to medium
 
 uniform float iTime;
 uniform vec3 iResolution;
