+++ Projects/ShaderToy/Shaders/GasGiant.fs
@@ -9,7 +9,7 @@
 // RAY V SPHERE (reinder): https://www.shadertoy.com/view/4tjGRh
 
 #version 300 es
-precision mediump float;       	// Set the default precision to medium
+precision highp float;       	// Set the default precision to medium
 
 uniform float iTime;
 uniform vec3 iResolution;
