+++ Projects/ShaderToy/Shaders/LakeDrops.fs
@@ -4,7 +4,7 @@
 //Horrible mod to non-cubemap texture!
 
 #version 300 es
-precision mediump float;       	// Set the default precision to medium
+precision highp float;       	// Set the default precision to medium
 
 uniform float iTime;
 uniform vec3 iResolution;
