+++ Projects/ShaderToy/Shaders/Mandelbrot.fs
@@ -16,7 +16,7 @@
 // More info here: http://www.iquilezles.org/www/articles/distancefractals/distancefractals.htm
 
 #version 300 es
-precision mediump float;       	// Set the default precision to medium
+precision highp float;       	// Set the default precision to medium
 
 uniform float iTime;
 uniform vec3 iResolution;
