+++ Projects/ShaderToy/Shaders/rainWaterRipple.fs
@@ -2,7 +2,7 @@
 //https://www.shadertoy.com/view/Mt33DH
 
 #version 300 es
-precision mediump float;       	// Set the default precision to medium
+precision highp float;       	// Set the default precision to medium
 
 uniform float iTime;
 uniform vec3 iResolution;
